+++ apps/client/src/pages/my-history/hooks/use-setlist-detail-mutation.ts
@@ -1,4 +1,4 @@
-import { useMutation, useSuspenseQuery } from '@tanstack/react-query';
+import { useMutation } from '@tanstack/react-query';
 
 import {
   deleteCancelEditSetList,
@@ -6,12 +6,6 @@ import {
   patchReorderSetList,
   postStartEditSetList,
 } from '@shared/apis/my-history/setlist-mutation';
-import { SETLIST_QUERY_OPTION } from '@shared/apis/my-history/setlist-queries';
-
-export const useSetListDetail = (setlistId: number) => {
-  const { data } = useSuspenseQuery(SETLIST_QUERY_OPTION.DETAIL(setlistId));
-  return { data };
-};
 
 export const useStartEditSetList = () => {
   return useMutation({
